Bugfix(msdf): Coloring of the first edge was not being done correctly.
If the edge started with the first segment in the contour, then it would
immediately switch color, ignoring the special case first color. If the
edge is split between the start and end segments of the contour, the two
parts would end up with different colors, resulting in a "fake" corner.
Both of these issues are now fixed.