Opt(render/opengl): Create a new vertex buffer for every batch
Turns out this significantly faster than reusing small buffers. Most
likely this is because we no longer have to wait for the GPU to finish
rendering if we end up reusing the same buffer (which we do most of the
time). In release mode, this improves `render_submit` from 3-4ms per
frame to 200us per frame.