Bugfix(font/msdf-gen): Correctly loop over all pairs of contours
Consider the following case. There are three contours: left, middle, and
right. The middle one overlaps the other two, but the left and right
ones are not overlapping. Kind of like this:

```
+---+   +---+
|   |   |   |
| +-------+ |
| | |   | | |
| +-------+ |
|   |   |   |
+---+   +---+
```

Furthermore, the contours are defined in this order: left, right,
middle. If this is the case, we will first detect two intersections
between the left contour and the middle one, and merge them. We would
then get something like this:

```
+---+   +---+
|   |   |   |
| +-+-----+ |
| | |   | | |
| +-+-----+ |
|   |   |   |
+---+   +---+
```

The T-shaped contour is now the left one because of the merging. When we
now check the right contour for overlap, we would previously never
compare it to the left one, because it is earlier in the list. As this
example illustrates, this is incorrect and we should instead loop over
all contours (skipping so we don't compare against ourselves).