Bugfix(font/ttf): Guard against empty contours
I hadn't thought someone would ever create a font with empty contours in
it, but here we are. This is not a perfect solution and is an indication
that we need to look over the entire parser and make sure that it is
robust and can handle any case that is thrown at it. We should be able
to load any font file that is thrown at us. I should probably look into
fuzzing.