Dev(msdf-gen): Cache glyph indicies rather than codepoints
As all information about a glyph is stored per glyph index, and multiple
codepoints can map to the same glyph index, this means that the
information can be reused. So now, rather than generating ~1 million
missing glyphs, we generate one and reuse it for all codepoints that
aren't mapped to anything. Computing glyph indicies is relatively cheap
so this should be fine.