Dev(msdf-gen): Factored out the MSDF cache
I'm almost happy with how it works. Currently, we send pointers to the
work thread, which I would like to change to handles instead for
stability purposes. You can easily see how this goes wrong if someone
unloaded when the work thread is in the middle of processing a glyph.

Secondly, I would like the size of the glyph to be customizable per
glyph, and not a constant for the cache. This would allow for more
flexibility.

You would probably want to turn it into a LRU cache as well for when we
implement font unloading.