Dev(graphics/wayland): Added server side decorations with XDG decoration manager
Not every compositor implements this extension and it is unstable. The
alternative to have native window decorations is libdecor, which is in
version 0.2.2 as of writing. It also messes with your event loop and so
this is less intrusive and works on most compositors.

If libdecor reaches version 1.0 and becomes less intrusive, we might
switch to it.