Dev(graphics/wayland): Implemented key repeat
Unfortunately, this makes us not sleep properly anymore. We wakeup
because EGL receives events that we don't care about, causing us to
produce another frame, which causes EGL to wait for more events and so
on. I don't really know how to solve this, but it kinda sucks that we
produce a bunch of frames when we don't need to.