Dev(font/ttf): Maintain the hierarchy of compound glyphs
This will allow us to visualize more information later on, such as
showing the different components, and the instructions for each
component.

This change also removes a limitation in the parser. Previously it
needed to know the maximum number of points in any glyph ahead of time.
This is documented in the maxp table. We no longer need this info, and
can thus read files that are internally inconsistent without failing. We
now instead have the issue that we can get stuck in infinite recursion.
This should be fixed later.